tutorials/vlm-workshop/requirements.txt
pandas
numpy
tqdm
pillow
datasets
matplotlib
seaborn
scikit-learn
ipython
jupyter
nbconvert
openai
python-dotenv
sentence-transformers
faiss-cpu
requests
